:alien: admin reviews api 변경에 따른 구조 변경 (related to #20)
